+++ app/Http/Controllers/PlacefreatureController.php
@@ -3,16 +3,18 @@
 namespace App\Http\Controllers;
 
 use App\Placefreature;
+use Illuminate\Http\RedirectResponse;
 use Illuminate\Http\Request;
+use Illuminate\Http\Response;
 
 class PlacefreatureController extends Controller
 {
     /**
      * Display a listing of the resource.
      *
-     * @return \Illuminate\Http\Response
+     * @return Response
      */
-    public function index()
+    public function index(): Response
     {
         $freatures = Placefreature::query()->orderBy('name')->get();
         return response(view('placefreature.index', compact('freatures')), 200);
@@ -21,9 +23,9 @@ class PlacefreatureController extends Controller
     /**
      * Show the form for creating a new resource.
      *
-     * @return \Illuminate\Http\Response
+     * @return Response
      */
-    public function create()
+    public function create(): Response
     {
         return response(view('placefreature.create'), 200);
     }
@@ -31,10 +33,10 @@ class PlacefreatureController extends Controller
     /**
      * Store a newly created resource in storage.
      *
-     * @param  \Illuminate\Http\Request  $request
-     * @return \Illuminate\Http\Response
+     * @param Request $request
+     * @return Response
      */
-    public function store(Request $request)
+    public function store(Request $request): Response
     {
         if (empty($request)) {
             return response('Formulário vazio');
@@ -48,21 +50,22 @@ class PlacefreatureController extends Controller
      * Display the specified resource.
      *
      * @param  int  $id
-     * @return \Illuminate\Http\Response
+     * @return Response
+     *
+     * public function show($id)
+    * {
+        * $freature = Placefreature::find($id);
+        * return response(view('placefreature.one', compact('freature')));
+    * }
      */
-    public function show($id)
-    {
-        $freature = Placefreature::find($id);
-        return response(view('placefreature.one', compact('freature')));
-    }
 
     /**
      * Show the form for editing the specified resource.
      *
      * @param  int  $id
-     * @return \Illuminate\Http\Response
+     * @return Response
      */
-    public function edit($id)
+    public function edit(int $id): Response
     {
         $freature = Placefreature::find($id);
         return response(view('placefreature.update', compact('freature')));
@@ -71,27 +74,32 @@ class PlacefreatureController extends Controller
     /**
      * Update the specified resource in storage.
      *
-     * @param  \Illuminate\Http\Request  $request
+     * @param Request $request
      * @param  int  $id
-     * @return \Illuminate\Http\Response
+     * @return Response
      */
-    public function update(Request $request, $id)
+    public function update(Request $request, int $id): Response
     {
         $freature = Placefreature::find($id);
         $freature->update($request->all());
 
-        return response(redirect()->route('show-placefreature', $freature->id));
+        return response(redirect()->route('index-placefreature', $freature->id));
     }
 
     /**
      * Remove the specified resource from storage.
      *
      * @param  int  $id
-     * @return \Illuminate\Http\Response
+     * @return RedirectResponse|Response
      */
-    public function destroy($id)
+    public function destroy(int $id): Response
     {
         $freature = Placefreature::find($id);
+
+        if ($freature->occurrences()->exists()) {
+            return back()->withErrors(['error' => 'Há ocorrências utilizando esse elemento']);
+        }
+
         $freature->delete();
 
         return response(redirect(route('index-placefreature')));
