+++ app/Http/Controllers/PlaceuseController.php
@@ -3,16 +3,18 @@
 namespace App\Http\Controllers;
 
 use App\Placeuse;
+use Illuminate\Http\RedirectResponse;
 use Illuminate\Http\Request;
+use Illuminate\Http\Response;
 
 class PlaceuseController extends Controller
 {
     /**
      * Display a listing of the resource.
      *
-     * @return \Illuminate\Http\Response
+     * @return Response
      */
-    public function index()
+    public function index(): Response
     {
         $uses = Placeuse::query()->orderBy('name')->get();
         return response(view('placeuse.index', compact('uses')), 200);
@@ -21,9 +23,9 @@ class PlaceuseController extends Controller
     /**
      * Show the form for creating a new resource.
      *
-     * @return \Illuminate\Http\Response
+     * @return Response
      */
-    public function create()
+    public function create(): Response
     {
         return response(view('placeuse.create'), 200);
     }
@@ -31,10 +33,10 @@ class PlaceuseController extends Controller
     /**
      * Store a newly created resource in storage.
      *
-     * @param  \Illuminate\Http\Request  $request
-     * @return \Illuminate\Http\Response
+     * @param Request $request
+     * @return Response
      */
-    public function store(Request $request)
+    public function store(Request $request): Response
     {
         if (empty($request)){
             return response('Formulário vazio');
@@ -48,21 +50,22 @@ class PlaceuseController extends Controller
      * Display the specified resource.
      *
      * @param  int  $id
-     * @return \Illuminate\Http\Response
+     * @return Response
+     *
+     * public function show($id)
+    * {
+        * $use = Placeuse::find($id);
+        * return response(view('placeuse.one', compact('use')));
+    * }
      */
-    public function show($id)
-    {
-        $use = Placeuse::find($id);
-        return response(view('placeuse.one', compact('use')));
-    }
 
     /**
      * Show the form for editing the specified resource.
      *
      * @param  int  $id
-     * @return \Illuminate\Http\Response
+     * @return Response
      */
-    public function edit($id)
+    public function edit(int $id): Response
     {
         $use = Placeuse::find($id);
         return response(view("placeuse.update", compact('use')));
@@ -71,27 +74,32 @@ class PlaceuseController extends Controller
     /**
      * Update the specified resource in storage.
      *
-     * @param  \Illuminate\Http\Request  $request
+     * @param Request $request
      * @param  int  $id
-     * @return \Illuminate\Http\Response
+     * @return Response
      */
-    public function update(Request $request, $id)
+    public function update(Request $request, int $id): Response
     {
         $use = Placeuse::find($id);
         $use->update($request->all());
 
-        return response(redirect()->route('show-placeuse', $use->id));
+        return response(redirect()->route('index-placeuse', $use->id));
     }
 
     /**
      * Remove the specified resource from storage.
      *
      * @param  int  $id
-     * @return \Illuminate\Http\Response
+     * @return RedirectResponse|Response
      */
-    public function destroy($id)
+    public function destroy(int $id): Response
     {
         $use = Placeuse::find($id);
+
+        if ($use->occurrences()->exists()) {
+            return back()->withErrors(['error' => 'Há ocorrências utilizando esse elemento']);
+        }
+
         $use->delete();
 
         return response(redirect(route('index-placeuse')));
