+++ app/Http/Controllers/ProblemController.php
@@ -3,16 +3,18 @@
 namespace App\Http\Controllers;
 
 use App\Problem;
+use Illuminate\Http\RedirectResponse;
 use Illuminate\Http\Request;
+use Illuminate\Http\Response;
 
 class ProblemController extends Controller
 {
     /**
      * Display a listing of the resource.
      *
-     * @return \Illuminate\Http\Response
+     * @return Response
      */
-    public function index()
+    public function index(): Response
     {
         $problems = Problem::query()->orderBy('name')->get();
         return response(view('problem.index', compact('problems')), 200);
@@ -21,9 +23,9 @@ class ProblemController extends Controller
     /**
      * Show the form for creating a new resource.
      *
-     * @return \Illuminate\Http\Response
+     * @return Response
      */
-    public function create()
+    public function create(): Response
     {
         return response(view('problem.create'), 200);
     }
@@ -31,10 +33,10 @@ class ProblemController extends Controller
     /**
      * Store a newly created resource in storage.
      *
-     * @param  \Illuminate\Http\Request  $request
-     * @return \Illuminate\Http\Response
+     * @param Request $request
+     * @return Response
      */
-    public function store(Request $request)
+    public function store(Request $request): Response
     {
         if (empty($request)) {
             return response('Formulário vazio');
@@ -48,9 +50,9 @@ class ProblemController extends Controller
      * Display the specified resource.
      *
      * @param  int  $id
-     * @return \Illuminate\Http\Response
+     * @return Response
      */
-    public function show($id)
+    public function show(int $id): Response
     {
         $problem = Problem::find($id);
         return response(view('problem.one', compact('problem')));
@@ -60,9 +62,9 @@ class ProblemController extends Controller
      * Show the form for editing the specified resource.
      *
      * @param  int  $id
-     * @return \Illuminate\Http\Response
+     * @return Response
      */
-    public function edit($id)
+    public function edit(int $id): Response
     {
         $problem = Problem::find($id);
         return response(view('problem.update', compact('problem')));
@@ -71,11 +73,11 @@ class ProblemController extends Controller
     /**
      * Update the specified resource in storage.
      *
-     * @param  \Illuminate\Http\Request  $request
+     * @param Request $request
      * @param  int  $id
-     * @return \Illuminate\Http\Response
+     * @return Response
      */
-    public function update(Request $request, $id)
+    public function update(Request $request, int $id): Response
     {
         $problem = Problem::find($id);
         $problem->update($request->all());
@@ -87,11 +89,16 @@ class ProblemController extends Controller
      * Remove the specified resource from storage.
      *
      * @param  int  $id
-     * @return \Illuminate\Http\Response
+     * @return RedirectResponse|Response
      */
-    public function destroy($id)
+    public function destroy(int $id)
     {
         $problem = Problem::find($id);
+
+        if ($problem->occurrences()->exists()) {
+            return back()->withErrors(['error' => 'Há vítimas utilizando esse elemento']);
+        }
+
         $problem->delete();
 
         return response(redirect(route('index-problem')));
