+++ app/Http/Controllers/RescuerController.php
@@ -3,16 +3,18 @@
 namespace App\Http\Controllers;
 
 use App\Rescuer;
+use Illuminate\Http\RedirectResponse;
 use Illuminate\Http\Request;
+use Illuminate\Http\Response;
 
 class RescuerController extends Controller
 {
     /**
      * Display a listing of the resource.
      *
-     * @return \Illuminate\Http\Response
+     * @return Response
      */
-    public function index()
+    public function index(): Response
     {
         $rescuers = Rescuer::query()->orderBy('name')->get();
         return response(view('rescuer.index', compact('rescuers')), 200);
@@ -21,9 +23,9 @@ class RescuerController extends Controller
     /**
      * Show the form for creating a new resource.
      *
-     * @return \Illuminate\Http\Response
+     * @return Response
      */
-    public function create()
+    public function create(): Response
     {
         return response(view('rescuer.create'), 200);
     }
@@ -31,10 +33,10 @@ class RescuerController extends Controller
     /**
      * Store a newly created resource in storage.
      *
-     * @param  \Illuminate\Http\Request  $request
-     * @return \Illuminate\Http\Response
+     * @param Request $request
+     * @return Response
      */
-    public function store(Request $request)
+    public function store(Request $request): Response
     {
         if (empty($request)) {
             return response('Formulário vazio');
@@ -48,21 +50,22 @@ class RescuerController extends Controller
      * Display the specified resource.
      *
      * @param  int  $id
-     * @return \Illuminate\Http\Response
-     */
+     * @return Response
+
     public function show($id)
     {
         $rescuer = Rescuer::find($id);
         return response(view('rescuer.one', compact('rescuer')));
     }
+     */
 
     /**
      * Show the form for editing the specified resource.
      *
      * @param  int  $id
-     * @return \Illuminate\Http\Response
+     * @return Response
      */
-    public function edit($id)
+    public function edit(int $id): Response
     {
         $rescuer = Rescuer::find($id);
         return response(view('rescuer.update', compact('rescuer')));
@@ -71,27 +74,32 @@ class RescuerController extends Controller
     /**
      * Update the specified resource in storage.
      *
-     * @param  \Illuminate\Http\Request  $request
+     * @param Request $request
      * @param  int  $id
-     * @return \Illuminate\Http\Response
+     * @return Response
      */
-    public function update(Request $request, $id)
+    public function update(Request $request, int $id): Response
     {
         $rescuer = Rescuer::find($id);
         $rescuer->update($request->all());
 
-        return response(redirect()->route('show-rescuer', $rescuer->id));
+        return response(redirect()->route('index-rescuer', $rescuer->id));
     }
 
     /**
      * Remove the specified resource from storage.
      *
      * @param  int  $id
-     * @return \Illuminate\Http\Response
+     * @return RedirectResponse|Response
      */
-    public function destroy($id)
+    public function destroy( int $id)
     {
         $rescuer = Rescuer::find($id);
+
+        if ($rescuer->victims()->exists()) {
+            return back()->withErrors(['error' => 'Há vítimas utilizando esse elemento']);
+        }
+
         $rescuer->delete();
 
         return response(redirect(route('index-rescuer')));
